+++ eck-autocomplete/utils/baseComponent.ts
@@ -9,6 +9,6 @@ export abstract class BaseComponent extends HTMLElement {
   injectCSS(css: string): void {
     const cssTemplate = document.createElement('template');
     cssTemplate.innerHTML = `<style>${css}</style>`;
-    this.shadowRoot!.appendChild(cssTemplate.content.cloneNode(true));
+    this.shadowRoot?.appendChild(cssTemplate.content.cloneNode(true));
   }
 }